new script to start tracking files

+++ symlink.sh
@@ -13,7 +13,7 @@ for f in $files; do
 
     # if the target file doesn't exist _and_ is a symlink (aka broken symlink),
     # remove it
-    if [ ! -f $target ] && [ -h $target ]; then
+    if [ ! -f $target ] && [ -L $target ]; then
         rm $target
     fi
 
@@ -22,9 +22,13 @@ for f in $files; do
         echo ln -s $(pwd)/$f $target
         ln -s $(pwd)/$f $target
     else
-        echo "$target already exists"
+        echo "[$target] already linked, skipping"
     fi
 done
 
-# .local/bin is an exception: it _should_ link back to this repo
-ln -s $(pwd)/home/t/.local/bin /home/t/.local/bin
+# .local/bin is an exception: it _should_ link back to this repo, so I don't
+# have to add each new script
+if [ ! -f /home/t/.local/bin ] && [ ! -L /home/t/.local/bin ]; then
+    echo ln -s $(pwd)/home/t/.local/bin /home/t/.local/bin
+    ln -s $(pwd)/home/t/.local/bin /home/t/.local/bin
+fi
